wallet manager is an app now
