+++ src/apps/wallets/__init__.py
@@ -1,0 +1,1 @@
+from .manager import WalletManager as App
