+++ src/apps/wallets/commands.py
@@ -1,0 +1,4 @@
+"""Commands coming from the GUI"""
+
+DELETE = 255
+EDIT = 1
