+++ src/gui/screens/__init__.py
@@ -6,6 +6,6 @@ from .qralert import QRAlert
 from .progress import Progress
 from .input import PinScreen, InputScreen, DerivationScreen
 from .mnemonic import MnemonicScreen, NewMnemonicScreen, RecoverMnemonicScreen
-from .bitcoin import XPubScreen, WalletScreen, ConfirmWalletScreen
+from .xpub import XPubScreen
 from .transaction import TransactionScreen
 from .settings import DevSettings
