+++ src/hosts/__init__.py
@@ -1,4 +1,4 @@
-from .core import HostError
+from .core import HostError, Host
 from .qr   import QRHost
 from .sd   import SDHost
 from .usb  import USBHost
