+++ src/hosts/qr.py
@@ -2,8 +2,6 @@ from .core import Host, HostError
 import pyb, time, asyncio
 from platform import simulator
 from io import BytesIO
-from binascii import b2a_base64, a2b_base64
-from bitcoin.psbt import PSBT, InputScope
 
 QRSCANNER_TRIGGER = "D5"
 # OK response from scanner
@@ -297,7 +295,7 @@ class QRHost(Host):
                 "Scanning...", 
                 "Point scanner to the QR code")
         data = await self.scan()
-        return self.parse(data)
+        return BytesIO(data)
 
     async def send_data(self, stream, meta):
         response = stream.read().decode()
@@ -307,28 +305,10 @@ class QRHost(Host):
             title = meta["title"]
         if "note" in meta:
             note = meta["note"]
-        await self.manager.gui.qr_alert(title, response, response, note=note)
-
-    def parse(self, data):
-        # TODO: refactor with ramfiles
-        if data is None:
-            return None, None
-        # probably base64-encoded PSBT
-        if data[:4] == b"cHNi":
-            try:
-                psbt = a2b_base64(data)
-                if psbt[:5] != b"psbt\xff":
-                    raise HostError("Not a PSBT")
-                return self.SIGN_PSBT, BytesIO(psbt)
-            except:
-                pass
-        # probably wallet descriptor
-        if b"&" in data and "?" not in data:
-            return self.ADD_WALLET, BytesIO(data)
-        # probably verifying address
-        if data.startswith(b"bitcoin:") or b"index=" in data:
-            return self.VERIFY_ADDRESS, BytesIO(data)
-        return self.UNKNOWN, BytesIO(data)
+        msg = response
+        if "message" in meta:
+            msg = meta["message"]
+        await self.manager.gui.qr_alert(title, msg, response, note=note, qr_width=480)
 
     @property
     def in_progress(self):
@@ -346,13 +326,3 @@ class QRHost(Host):
         if not self.animated:
             return 0
         return [len(part)>0 for part in self.parts]
-    
-    async def send_psbt(self, psbt):
-        # we only keep partial sigs in the psbt
-        out_psbt = PSBT(psbt.tx)
-        for i, inp in enumerate(psbt.inputs):
-            out_psbt.inputs[i].partial_sigs = inp.partial_sigs
-        txt = b2a_base64(out_psbt.serialize()).decode().strip("\n")
-        if self.manager is not None:
-            await self.manager.gui.qr_alert("Transaction is signed!", 
-                        "Scan this QR code with your wallet", txt, qr_width=480)
