+++ src/main.py
@@ -2,7 +2,6 @@ from specter import Specter
 from gui.specter import SpecterGUI
 from keystore import FlashKeyStore
 from hosts import SDHost, QRHost, USBHost
-from wallets import WalletManager
 import platform
 
 def main():
@@ -35,15 +34,10 @@ def main():
             applications.append(app)
         else:
             print("Failed loading app:", modname)
-    # define WalletManager, requires keystore
-    # to authenticate wallet files
-    wallets_path = platform.fpath("/qspi/wallets")
-    wallet_manager = WalletManager(wallets_path)
 
     # make Specter instance
     settings_path = platform.fpath("/flash")
     specter = Specter(gui=gui, 
-                      wallet_manager=wallet_manager,
                       keystore=keystore,
                       hosts=hosts,
                       apps=applications,
