+++ src/specter.py
@@ -5,13 +5,13 @@ import asyncio
 
 from keystore import FlashKeyStore
 from platform import CriticalErrorWipeImmediately, set_usb_mode, reboot
-from hosts import HostError
+from hosts import Host, HostError
+from app import BaseApp
 from bitcoin import ec, psbt, bip32, bip39, script
 from bitcoin.psbt import PSBT
 from bitcoin.networks import NETWORKS
 # small helper functions
 from helpers import gen_mnemonic
-from gui.commands import EDIT, DELETE
 from errors import BaseError
 
 class SpecterError(BaseError):
@@ -22,11 +22,10 @@ class Specter:
     Call .start() method to register in the event loop
     It will then call the .setup() and .main() functions to display the GUI
     """
-    def __init__(self, gui, wallet_manager, keystore, hosts, apps, settings_path):
+    def __init__(self, gui, keystore, hosts, apps, settings_path):
         self.hosts = hosts
         self.keystore = keystore
         self.gui = gui
-        self.wallet_manager = wallet_manager
         self.path = settings_path
         self.load_network(self.path)
         self.current_menu = self.initmenu
@@ -51,8 +50,6 @@ class Specter:
         try:
             raise exception
         except CriticalErrorWipeImmediately as e:
-            # wipe all wallets
-            self.wallet_manager.wipe()
             # wipe all apps
             for app in self.apps:
                 app.wipe()
@@ -147,7 +144,6 @@ class Specter:
             if mnemonic is not None:
                 # load keys using mnemonic and empty password
                 self.keystore.set_mnemonic(mnemonic.strip(),"")
-                self.wallet_manager.init(self.keystore, self.network)
                 for app in self.apps:
                     app.init(self.keystore, self.network)
                 return self.mainmenu
@@ -157,7 +153,6 @@ class Specter:
             if mnemonic is not None:
                 # load keys using mnemonic and empty password
                 self.keystore.set_mnemonic(mnemonic,"")
-                self.wallet_manager.init(self.keystore, self.network)
                 for app in self.apps:
                     app.init(self.keystore, self.network)
                 self.current_menu = self.mainmenu
@@ -165,7 +160,6 @@ class Specter:
         elif menuitem == 2:
             self.keystore.load_mnemonic()
             # await self.gui.alert("Success!", "Key is loaded from flash!")
-            self.wallet_manager.init(self.keystore, self.network)
             for app in self.apps:
                 app.init(self.keystore, self.network)
             return self.mainmenu
@@ -202,13 +196,18 @@ class Specter:
         host_buttons = [
             (host, host.button) for host in self.hosts if host.button is not None
         ]
+        # buttons defined by app classes
+        app_buttons = [
+            (app, app.button) for app in self.apps if app.button is not None
+        ]
         # for every button we use an ID
         # to avoid mistakes when editing strings
         # If ID is None - it is a section title, not a button
         buttons = [
             # id, text
+            (None, "Applications"),
+        ] + app_buttons + [
             (None, "Key management"),
-            (0, "Wallets"),
             (1, "Master public keys"),
             (None, "Communication"),
         ] + host_buttons + [
@@ -221,11 +220,8 @@ class Specter:
         menuitem = await self.gui.menu(buttons)
 
         # process the menu button:
-        # wallets
-        if menuitem == 0:
-            return await self.show_wallets()
         # xpubs
-        elif menuitem == 1:
+        if menuitem == 1:
             return await self.show_master_keys()
         # lock device
         elif menuitem == 2:
@@ -236,40 +232,17 @@ class Specter:
             await self.select_network()
         elif menuitem == 4:
             return await self.settingsmenu()
-        # if it's a host
-        elif hasattr(menuitem, 'get_data'):
-            host = menuitem
-            cmd, stream = await host.get_data()
-            if cmd == host.SIGN_PSBT:
-                res = await self.sign_psbt(stream)
-                if res is not None:
-                    await host.send_psbt(res)
-            elif cmd == host.ADD_WALLET:
-                # read content, it's small
-                desc = stream.read().decode()
-                w = self.wallet_manager.parse_wallet(desc)
-                res = await self.confirm_new_wallet(w)
-                if res:
-                    self.wallet_manager.add_wallet(w)
-                    return await self.show_wallets()
-            elif cmd == host.VERIFY_ADDRESS:
-                data = stream.read().decode().replace("bitcoin:","")
-                # should be of the form addr?index=N or similar
-                if "index=" not in data or "?" not in data:
-                    raise SpecterError("Can't verify address with unknown index")
-                addr, rest = data.split("?")
-                args = rest.split("&")
-                idx = None
-                for arg in args:
-                    if arg.startswith("index="):
-                        idx = int(arg[6:])
-                        break
-                w = self.wallet_manager.find_wallet_from_address(addr, idx)
-                await self.gui.show_wallet(w, self.network, idx)
-            # probably user cancelled
-            elif cmd == None:
+        elif isinstance(menuitem, BaseApp) and hasattr(menuitem, 'menu'):
+            app = menuitem
+            # stay in this menu while something is returned
+            while await app.menu(self.gui.show_screen()):
                 pass
-            else:
+        # if it's a host
+        elif isinstance(menuitem, Host) and hasattr(menuitem, 'get_data'):
+            host = menuitem
+            stream = await host.get_data()
+            # probably user cancelled
+            if stream is not None:
                 # check against all apps
                 res = await self.process_host_request(stream, popup=False)
                 if res is not None:
@@ -277,18 +250,6 @@ class Specter:
         else:
             print(menuitem)
             raise SpecterError("Not implemented")
-
-    async def sign_psbt(self, stream, remote=False):
-        psbt = PSBT.read_from(stream)
-        wallet, meta = self.wallet_manager.parse_psbt(psbt=psbt)
-        res = await self.gui.confirm_transaction(wallet.name, meta, remote=remote)
-        if res:
-            # fill derivation paths from proprietary fields
-            wallet.update_gaps(psbt=psbt)
-            wallet.save(self.keystore)
-            psbt = wallet.fill_psbt(psbt, self.keystore.fingerprint)
-            self.keystore.sign_psbt(psbt)
-            return psbt
 
     async def process_host_request(self, stream, popup=True):
         matching_apps = []
@@ -300,14 +261,10 @@ class Specter:
         if len(matching_apps) == 0:
             raise HostError("Host command is not recognized")
         # TODO: if more than one - ask which one to use
+        if len(matching_apps) > 1:
+            raise HostError("Not sure what app to use... There are %d" % len(matching_apps))
         stream.seek(0)
         return await matching_apps[0].process_host_command(stream, self.gui.show_screen(popup))
-
-    async def confirm_new_wallet(self, w):
-        keys = [{"key": k, "mine": self.keystore.owns(k)} for k in w.get_keys()]
-        if not any([k["mine"] for k in keys]):
-            raise SpecterError("None of the keys belong to the device")
-        return await self.gui.confirm_new_wallet(w.name, w.policy, keys)
 
     async def select_network(self):
         # dict is unordered unfortunately, so we need to use hardcoded arr
@@ -328,7 +285,6 @@ class Specter:
             f.write(net)
         if self.keystore.is_ready:
             # load wallets for this network
-            self.wallet_manager.init(self.keystore, self.network)
             for app in self.apps:
                 app.init(self.keystore, self.network)
 
@@ -348,7 +304,7 @@ class Specter:
             # id, text
             (None, "Key management"),
             (0, "Reckless"),
-            (2, "Enter a bip39 password"),
+            (2, "Enter password"),
             (None, "Security"), # delimiter
             (3, "Change PIN code"),
             (4, "Developer & USB"),
@@ -367,7 +323,6 @@ class Specter:
             if pwd is None:
                 return self.mainmenu
             self.keystore.set_mnemonic(password=pwd)
-            self.wallet_manager.init(self.keystore, self.network)
             for app in self.apps:
                 app.init(self.keystore, self.network)
         elif menuitem == 3:
@@ -420,29 +375,6 @@ class Specter:
             await self.gui.show_mnemonic(self.keystore.mnemonic)
         else:
             raise SpecterError("Invalid menu")
-
-    async def show_wallets(self):
-        buttons = [(None, "Your wallets")]
-        buttons += [(i, w.name) for i, w in enumerate(self.wallet_manager.wallets)]
-        menuitem = await self.gui.menu(buttons, last=(255,None))
-        if menuitem == 255:
-            return self.mainmenu
-        else:
-            w = self.wallet_manager.wallets[menuitem]
-            # pass wallet and network
-            cmd = await self.gui.show_wallet(w, self.network, idx=w.unused_recv)
-            if cmd == DELETE:
-                conf = await self.gui.prompt("Delete wallet?", "You are deleting wallet \"%s\".\nAre you sure you want to do it?" % w.name)
-                if conf:
-                    self.wallet_manager.delete_wallet(w)
-                return self.show_wallets
-            elif cmd == EDIT:
-                name = await self.gui.get_input(title="Enter new wallet name", 
-                            note="", suggestion=w.name)
-                if name is not None and name != w.name and name != "":
-                    w.name = name
-                    w.save(self.keystore)
-                return self.show_wallets
 
     async def lock(self):
         # lock the keystore
@@ -512,55 +444,6 @@ class Specter:
             raise SpecterError("Device is locked")
         xpub = self.keystore.get_xpub(bip32.path_to_str(path))
         return xpub
-
-    async def showaddr(self, paths:list, script_type:str, redeem_script=None) -> str:
-        if redeem_script is not None:
-            redeem_script = script.Script(unhexlify(redeem_script))
-        # first check if we have corresponding wallet:
-        # - just take last 2 indexes of the derivation
-        # and see if redeem script matches
-        address = None
-        if redeem_script is not None:
-            if script_type == b"wsh":
-                address = script.p2wsh(redeem_script).address(NETWORKS[self.network])
-            elif script_type == b"sh-wsh":
-                address = script.p2sh(
-                            script.p2wsh(redeem_script)
-                          ).address(NETWORKS[self.network])
-            else:
-                raise HostError("Unsupported script type: %s" % script_type)
-        # in our wallets every key 
-        # has the same two last indexes for derivation
-        path = paths[0]
-        if not path.startswith(b"m/"):
-            path = b"m"+path[8:]
-        derivation = bip32.parse_path(path.decode())
-
-        # if not multisig:
-        if address is None and len(paths)==1:
-            pub = self.keystore.get_xpub(derivation)
-            if script_type == b"wpkh":
-                address = script.p2wpkh(pub).address(NETWORKS[self.network])
-            elif script_type == b"sh-wpkh":
-                address = script.p2sh(
-                            script.p2wpkh(pub).address(NETWORKS[self.network])
-                          )
-            else:
-                raise HostError("Unsupported script type: %s" % script_type)
-
-        if len(derivation) >= 2:
-            derivation = derivation[-2:]
-        else:
-            raise HostError("Invalid derivation")
-        if address is None:
-            raise HostError("Can't derive address. Provide redeem script.")
-        try:
-            change = bool(derivation[0])
-            w = self.wallet_manager.find_wallet_from_address(address, derivation[1], change=change)
-        except BaseError as e:
-            raise HostError("%s" % e)
-        await self.gui.show_wallet(w, self.network, derivation[1], change=change, remote=True)
-        return address
 
     async def show_master_keys(self, show_all=False):
         net = NETWORKS[self.network]
