--- src/wallets/__init__.py
@@ -1,3 +1,0 @@
-from .manager import WalletManager
-from .wallet import Wallet, WalletError
-from .scripts import ScriptError
