+++ test/run_tests.py
@@ -2,6 +2,7 @@ import sys
 sys.path.append('../src')
 sys.path.append('../f469-disco/lib')
 sys.path.append('../f469-disco/lib_unixport')
+sys.path.append('../f469-disco/usermods/udisplay_f469/display_unixport')
 sys.path.append('../f469-disco/tests')
 
 import unittest
